turn style.css into a css page
Before, it was written as html but was being run as css. Now is being run as css. You can include all of the css for all the site, or certain parts here, link it to all of your sites by using a style tag with the src property, and you can delete it from each page. This will make the code neater (and may reduce loading time when navigating to different pages on your site). Also changed the header height, before, the navigation buttons where hidden as white on white.